+++ src/main/java/com/challenge/Allocation/repository/RoomRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository extends JpaRepository<Room, Long> {
 
-    Room findByNumber(int numero);
+    Room findByNumber(Long numero);
 
 }
